Add scripts folder

+++ .local/bin/lock.sh
@@ -1,8 +1,21 @@
 #!/bin/bash
 
-swayidle \
-    timeout 5 'swaymsg "output * dpms off"' \
-    resume 'swaymsg "output * dpms on"' &
+set -e
+
+if pidof swaylock > /dev/null; then
+    echo "error: already locked?"
+    exit 1
+fi
+
+cleanup() {
+    kill %%
+}
+
+swayidle timeout 10 'swaymsg "output * dpms off"' \
+         resume 'swaymsg "output * dpms on"' &
+
+# Be sure to kill swayidle.
+trap 'cleanup' SIGINT SIGTERM QUIT
 
 swaylock
 kill %%
